refactor: Extract BrowserTitle and HistoryIconButton components
