+++ packages/kit/src/views/Discovery/pages/components/BrowserTitle.tsx
@@ -1,0 +1,16 @@
+import { useIntl } from 'react-intl';
+
+import { SizableText } from '@onekeyhq/components';
+import { ETranslations } from '@onekeyhq/shared/src/locale';
+
+export function BrowserTitle() {
+  const intl = useIntl();
+
+  return (
+    <SizableText size="$headingLg" color="$text">
+      {intl.formatMessage({ id: ETranslations.global_browser })}
+    </SizableText>
+  );
+}
+
+export default BrowserTitle;
